# EinStealthBackend/src/api/docs/swagger.yaml
definitions:
  controller.GetSpaceTimes.JsonResponse:
    properties:
      altitude:
        type: number
      latitude:
        type: number
      longitude:
        type: number
      objId:
        description: '人: player_id, 罠: player_id*(-1)'
        type: integer
    type: object
  controller.PostSpaceTimes.JsonRequest:
    properties:
      altitude:
        type: number
      latitude:
        type: number
      longitude:
        type: number
      obj_id:
        type: integer
      time:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/players/{id}/status/{status}:
    post:
      consumes:
      - application/json
      parameters:
      - description: player id
        in: path
        name: id
        required: true
        type: integer
      - description: player status
        in: path
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: プレイヤーの状態を更新するAPI
  /api/v1/spacetimes:
    get:
      consumes:
      - application/json
      parameters:
      - description: time
        in: query
        name: time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.GetSpaceTimes.JsonResponse'
            type: array
      summary: 指定された時刻の(位置, 罠or人, プレイヤーid, 状態)を取得する
    post:
      consumes:
      - application/json
      parameters:
      - description: request json
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.PostSpaceTimes.JsonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 指定された時刻の(位置, 罠or人, プレイヤーid, 状態)を格納するAPI
  /api/v1/room:
    get:
      consumes:
      - application/json
      parameters:
      - description: secret_words
        in: query
        name: secret_words
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.GetRoom.JsonResponse'
            type: array
      summary: 指定された部屋の(合言葉、状態)を返す
    post:
      consumes:
      - application/json
      parameters:
      - description: request json
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.PostRoom.JsonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: 指定された部屋の(合言葉、状態)を返す
  /ping:
    get:
      consumes:
      - application/json
      description: ping pong
      responses:
        "200":
          description: pong
          schema:
            type: string
      summary: test api
swagger: "2.0"